update pinned pico-sdk version to 2.1.1
Signed-off-by: Peter Maxwell Warasila <madmaxwell@soundcomesout.com

This updated release contains many improvements and (more importantly)
solves the CMake minimum required version issues.

The sleep functionality has been broken by this, but it does not crash
the program when the badge attempts to slee. This and other issues
surrounding sleep should be fixed as they are discovered rather than
blocking all development on machines with CMake 4.x.x.