fix screensaver breakage on hardware
Issuing a command to the display has two parts. First, a command is sent
to the display over the SPI bus. Then, the data for that command is sent
over the bus. One of the pins to the display toggles between the data
and command modes.

While the synchroniziation correctly waited for the previous SPI
transmission to finish before sending new dta, the driver did not wait
for the previous transmission to complete before toggling this pin. This
causes the dispaly to interpret the currently transmitted data as
commands leaving the display in an undefined state.

This is normally only possible if the application peforms an unexpected
set of operations (namely flushing the frame buffer to the display twice
in quick succession). However, it is possible for this to occur in the
disp_asset_saver() function on the last frame when going to sleep.
Because the image is flushed to the display by the screensaver and then
the frame buffer is cleared and flushed again before waiting for the
next frame time, the flush of the cleared frame buffer triggers this bug
when the image rectangle command is sent thereby toggling the
command/data pin while the previous flush is already in progress.

Signed-off-by: Peter Maxwell Warasila <madmaxwell@soundcomesout.com>