badgey: rename "badgey" to "RVAsec Quest" in the menu
It was named "badgey" in 2024, when the badge was styled to look like a planet
rover, so then it made sense that you were the badge rover, trying to get home.
Now, that theme doesn't make sense, and the app is re-styled such that you're
trying to get to RVAsec, and in order to do so you must gather the parts to
create a badge to enter the conference.  So the name "badgey" no longer makes
sense.  "RVAsec Quest" is the new name at least until a better idea comes
along.

Signed-off-by: Stephen M. Cameron <stephenmcameron@gmail.com>