readme: attempting to streamline the readme
The previous readme seemed well written.  Im attempting to
organize it by relevance to try to make it easier to mentally
parse for someone trying to get up and running.

Moving a lot of the vscode chatter to a different .md file.
Its optional and it seems most people that work on this dont
use vscode anyways.

Removing "TO DO"
Removing "Whats working."
Removing "Off-Target Unit Tests"
Removing "Generating Documentation"

Making the simulator slightly more prevelent in the README
because its what most people will be building/running most
of the time.