codec: configure microhpone as disabled
All of the plumbing for the microphone is present. However, the samples
from the codec do not appear to represent the input. For now, I've kept
all of the code/plumbing but have set all fo the microhpone/ADC hardware
to powered-off. All of the samples from the microphone will be 0.

Signed-off-by: Peter Maxwell Warasila <madmaxwell@soundcomesout.com>