screensaver and apps: use app->wake_up instead of tracking screensaver activity
I had just added an API for apps to detect if the screensaver was active,
but now I'm deleting it because we don't need it because we *already*
have such a thing in the app stack.  The badge apps callback function is
like this:

	int badge_app_callback(struct badge_app *app);

That app parameter has a "wake_up" field.  That wake_up field
will be true whenever the app is on the top of the app stack
due to a push_app() or pop_app(), which means it will be true
when the screensaver has just exited (i.e. pop_app()).

So we don't need to query whether the screensaver has run because
that's what app->wake_up is for, and app->wake_up solves it not
only for the screensaver, but for *any* app that might disturb
the screen (e.g. the hypothetical tamigotchi app, for example.)

Signed-off-by: Stephen M. Cameron <stephenmcameron@gmail.com>