badgemon: disable old badgemonsters
having both badgemonsters causes the binary to be larger (because the assets are in there twice).
I think the new one is close to beign ready and will likely get finished, but
the old badgemonsters is still around if we need to fall back to it.

disabling the old badge monsters.
adding the unlock monster function
adding the screensaver function