web: initial rework
I didnt receive any feedback on the web stuff.
I reworked the website to be more direct and extensible.
The goal is to reduce any reliance on 3rd party tools and deps.

Its a fairly simple site that operates kind of like a wiki.
Im also trying to improve the wasm sim experience.